changes in Save property click event
when we fill the details in the property form then created date and
created by is also filled

+++ RMS.Web/RMS.Models/Person.cs
@@ -14,8 +14,8 @@ namespace RMS.Models
         public string Address1  { get; set; }
         public string Address2 { get; set; }
         public Int64 PinNo { get; set; }
-        public int PhoneNo { get; set; }
-        public Int32 MobileNo { get; set; }
+        public Int64 PhoneNo { get; set; }
+        public Int64 MobileNo { get; set; }
         public string Email { get; set; }
         public int CreatedBy { get; set; }
         public DateTime CreatedDate{ get; set; }
